datahub-airflow/dags/vertical/team/entity/projects/product/country/auto_decisioning/auto_scoring/model/requirements.txt
# aiohappyeyeballs==2.4.0
# aiohttp==3.10.5
# aiosignal==1.3.1
# argparse
# async-timeout==4.0.3
# attrs==24.2.0
black
# cachetools==5.5.0
# certifi==2024.7.4
# charset-normalizer==3.3.2
# click==8.1.7
db-dtypes
# decorator==5.1.1
# filelock==3.15.4
flake8
FLAML==2.1.1
# frozenlist==1.4.1
fsspec
gcsfs
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-core
google-cloud-secret-manager
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
grpc-google-iam-v1
grpcio
grpcio-status
httplib2
idna
isort
Jinja2
joblib
jsonschema
jsonschema-specifications
lightgbm==4.5.0
# MarkupSafe==2.1.5
# msgpack==1.0.8
# multidict==6.0.5
# numpy==1.22.0
# oauthlib==3.2.2
onnx==1.16.2
onnxconverter_common==1.14.0
onnxmltools==1.12.0
onnxruntime==1.19.2
pandas==2.1.2
# proto-plus==1.22.3
# protobuf==4.25.0
# pyarrow==14.0.1
# pyasn1==0.6.0
# pyasn1_modules==0.4.0
# pyparsing==3.1.2
# python-dateutil==2.9.0.post0
# pytz==2024.1
PyYAML==6.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
scikit-learn==1.1.0
scipy==1.11.4
# six==1.16.0
sqlfluff
skl2onnx==1.17.0
# threadpoolctl==3.5.0
# uritemplate==4.1.1
# urllib3==2.2
xgboost==1.6.2
yamllint